## Summary

This PR implements comprehensive performance optimizations across multiple modules in the `est` library, with the most significant improvement being a **2.36x performance boost** for Unpin futures.

## Key Optimizations

### 🚀 Future Module - Major Performance Improvement
- **New**: `WithCancelSignalUnpin` structure for zero-cost abstraction over Unpin futures
- **New**: `with_cancel_signal_unpin()` method avoiding heap allocation
- **Benchmark**: 2.36x faster than boxed version (60ns vs 142ns)
- **Impact**: Significant performance improvement for async code using Unpin futures

### 🔧 Collections Module - Bug Fix + Optimization
- **Fixed**: HashMap/BTreeMap `replace_key` error handling order to maintain API contract
- **Added**: `#[inline]` hints for better compiler optimization
- **Impact**: Maintains correctness while improving hot path performance

### ⚡ Process Module - Reduced Overhead
- **Optimized**: Clone implementation using pattern matching to avoid redundant conversions
- **Improved**: Environment variable processing with batch operations
- **Impact**: Reduced CPU overhead for process command operations

### 🎯 Sync Module - Hot Path Optimization
- **Added**: `#[inline]` hint to `trigger()` method
- **Impact**: Better performance for critical synchronization operations

## Benchmark Results

```
with_cancel_signal_unpin: 60.29 ns
with_cancel_signal_boxed: 142.33 ns
Improvement: 2.36x faster
```

## Testing & Quality Assurance

- ✅ All 28 unit tests pass
- ✅ All 21 documentation tests pass (1 ignored)
- ✅ Clippy analysis clean (no warnings)
- ✅ Full backward compatibility maintained
- ✅ Comprehensive benchmarks added

## Documentation

- Added comprehensive performance optimization report (`PERFORMANCE_OPTIMIZATIONS.md`)
- All new APIs properly documented with examples
- Benchmark suite added for future performance tracking

## Compatibility

- **API**: Fully backward compatible
- **Rust Version**: Compatible with Rust 1.85.0+
- **Dependencies**: No new required dependencies
- **Features**: All feature flags work as expected

## Files Changed

- `src/collections.rs`: Fixed error handling order + inline optimizations
- `src/future.rs`: Added WithCancelSignalUnpin + inline optimizations
- `src/process.rs`: Optimized Clone implementation
- `src/sync/once.rs`: Added inline hint to trigger method
- `Cargo.toml`: Added criterion for benchmarking
- `benches/performance.rs`: Comprehensive benchmark suite
- `PERFORMANCE_OPTIMIZATIONS.md`: Detailed optimization report

These optimizations provide measurable performance improvements while maintaining the library's high code quality standards and full backward compatibility.